Refactor createThemeUtil
