+++ frontend/lib/src/mocks/mockTheme.ts
@@ -20,7 +20,7 @@ import { lightThemePrimitives } from "baseui"
 import { transparentize } from "color2k"
 
 import { ThemeConfig } from "~lib/theme"
-import { createBaseUiTheme } from "~lib/theme/createThemeUtil"
+import { createBaseUiTheme } from "~lib/theme/createBaseUiTheme"
 import { createEmotionColors } from "~lib/theme/getColors"
 import {
   breakpoints,
